chore: Extract common select component code (#21094)
